requirements.txt
pyside6
imagehash
jinja2
lxml
opencv-python-headless
openpyxl
numpy
pillow
python-dateutil
pywin32; sys_platform == 'win32'
zxing-cpp
